marketing-agency/requirements.txt
google-adk
google-genai
pydantic==2.10.6
python-dotenv==1.0.1
google-cloud-aiplatform
google-adk[eval]==1.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
black==25.1.0
absl-py==2.2.1
deprecated